Increase max payload size for hydration
The maximum URL size in Node 10+ is 8KB. We're setting a much lower
limit for hydration payloads (2KB). This commit bumps that limit to
get it much closer to the true upper bound.

Why not go for the true max of 6144 bytes? We add some additional
overhead when converting from 'hydrate|||...|||hydrate'. There's a
domain, a signature, other request headers, etc. 6000 pre-base64 is
a nice, round number to cut off at.