CRIT-371 perf(core): Optimize secret-scrubbing input in logger (#542)
* Rename truncate to truncateString

* Add truncateData function

This function will be used to truncate the data.input and data.output
before secret scrubbing and logging. It ensures that the JSON structure
of the passed in data remains intact and that the result, when
stringified, has a length <= the specified length.

* Refactor truncateData

Break up components into functions for legibility. Add docstrings.

* Improve truncateData testing

Use a fixture file for loading larger data without taking up too many
lines of code in the test file.